backend/requirements.txt
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
cssselect2==0.7.0
defusedxml==0.7.1
Django==3.2.16
django-cors-headers==4.3.0
django-extra-fields==3.0.2
django-filer==3.1.1
django-filter==23.3
django-log-viewer==1.1.7
django-polymorphic==3.1.0
django-templated-mail==1.1.1
django-username-email==2.5.4
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-filter==0.2.2
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-yasg==1.21.7
easy-thumbnails==2.8.5
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
lxml==4.9.3
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
Pillow==9.3.0
platformdirs==3.10.0
pluggy==0.13.1
psycopg2==2.9.7
py==1.11.0
pycparser==2.21
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
reportlab==4.0.7
requests==2.26.0
requests-oauthlib==1.3.1
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
webencodings==0.5.1
